Add gradle executable to build on shell Move data to common location and update file locations
